test(table): fix cyclic references
