+++ packages/table/src/wrapper/index.ts
@@ -1,4 +1,3 @@
 export * from './CellWrapper';
 export * from './CellEditModeWrapper';
 export * from './CellExpandWrapper';
-export * from './declarations';
